chore: #228 코드리뷰 반영
