+++ Spoony-iOS/Spoony-iOS/Source/Feature/Explore/Enums/ExploreSearchState.swift
@@ -1,0 +1,16 @@
+//
+//  ExploreSearchState.swift
+//  Spoony-iOS
+//
+//  Created by 최주리 on 4/20/25.
+//
+
+import Foundation
+
+enum ExploreSearchState {
+    case beforeSearch
+    case recentSearch
+    case searching
+    case searchResult
+    case noResult
+}
